requirements.txt
aiogram
aioredis
asyncpg
python-dotenv
sqlalchemy[async]